+++ Cargo.toml
@@ -68,7 +68,7 @@ deno_media_type = { version = "=0.2.8", features = ["module_specifier"] }
 deno_native_certs = "0.3.0"
 deno_npm = "=0.33.3"
 deno_package_json = { version = "=0.6.0", default-features = false }
-deno_path_util = "=0.3.2"
+deno_path_util = "=0.3.3"
 deno_semver = "=0.7.1"
 deno_task_shell = "=0.23.0"
 deno_terminal = "=0.2.2"
@@ -520,4 +520,5 @@ opt-level = 3
 opt-level = 3
 
 [patch.crates-io]
-sys_traits = { path = "../sys_traits" }
+deno_unsync = { path = "../deno_unsync" }
+deno_npm = { path = "../deno_npm" }
