+++ cli/lib/npm/mod.rs
@@ -2,10 +2,8 @@
 
 mod permission_checker;
 
-use std::path::Path;
 use std::sync::Arc;
 
-use deno_npm::resolution::ValidSerializedNpmResolutionSnapshot;
 use deno_resolver::npm::ByonmNpmResolver;
 use deno_resolver::npm::ManagedNpmResolverRc;
 use deno_resolver::npm::NpmResolver;
@@ -14,8 +12,6 @@ use deno_runtime::deno_process::NpmProcessStateProviderRc;
 pub use permission_checker::NpmRegistryReadPermissionChecker;
 pub use permission_checker::NpmRegistryReadPermissionCheckerMode;
 
-use crate::args::NpmProcessState;
-use crate::args::NpmProcessStateKind;
 use crate::sys::DenoLibSys;
 
 pub fn create_npm_process_state_provider<TSys: DenoLibSys>(
@@ -31,18 +27,6 @@ pub fn create_npm_process_state_provider<TSys: DenoLibSys>(
   }
 }
 
-pub fn npm_process_state(
-  snapshot: ValidSerializedNpmResolutionSnapshot,
-  node_modules_path: Option<&Path>,
-) -> String {
-  serde_json::to_string(&NpmProcessState {
-    kind: NpmProcessStateKind::Snapshot(snapshot.into_serialized()),
-    local_node_modules_path: node_modules_path
-      .map(|p| p.to_string_lossy().to_string()),
-  })
-  .unwrap()
-}
-
 #[derive(Debug)]
 pub struct ManagedNpmProcessStateProvider<TSys: DenoLibSys>(
   pub ManagedNpmResolverRc<TSys>,
@@ -52,7 +36,7 @@ impl<TSys: DenoLibSys> NpmProcessStateProvider
   for ManagedNpmProcessStateProvider<TSys>
 {
   fn get_npm_process_state(&self) -> String {
-    npm_process_state(
+    serialize_npm_process_state(
       self.0.resolution().serialized_valid_snapshot(),
       self.0.root_node_modules_path(),
     )
