+++ cli/npm/installer/mod.rs
@@ -39,274 +39,3 @@ mod common;
 mod global;
 mod local;
 mod resolution;
-
-#[derive(Debug)]
-pub struct NpmInstaller {
-  fs_installer: Arc<dyn NpmPackageFsInstaller>,
-  npm_install_deps_provider: Arc<NpmInstallDepsProvider>,
-  npm_resolution_initializer: Arc<NpmResolutionInitializer>,
-  npm_resolution_installer: Arc<NpmResolutionInstaller>,
-  maybe_lockfile: Option<Arc<CliLockfile>>,
-  npm_resolution: Arc<NpmResolutionCell>,
-  top_level_install_flag: AtomicFlag,
-  cached_reqs: tokio::sync::Mutex<FxHashSet<PackageReq>>,
-}
-
-impl NpmInstaller {
-  #[allow(clippy::too_many_arguments)]
-  pub fn new(
-    lifecycle_scripts_executor: Arc<dyn LifecycleScriptsExecutor>,
-    npm_cache: Arc<CliNpmCache>,
-    npm_install_deps_provider: Arc<NpmInstallDepsProvider>,
-    npm_registry_info_provider: Arc<
-      dyn deno_npm::registry::NpmRegistryApi + Send + Sync,
-    >,
-    npm_resolution: Arc<NpmResolutionCell>,
-    npm_resolution_initializer: Arc<NpmResolutionInitializer>,
-    npm_resolution_installer: Arc<NpmResolutionInstaller>,
-    progress_bar: &ProgressBar,
-    sys: CliSys,
-    tarball_cache: Arc<CliNpmTarballCache>,
-    maybe_lockfile: Option<Arc<CliLockfile>>,
-    maybe_node_modules_path: Option<PathBuf>,
-    lifecycle_scripts: LifecycleScriptsConfig,
-    system_info: NpmSystemInfo,
-    workspace_patch_packages: Arc<WorkspaceNpmPatchPackages>,
-  ) -> Self {
-    let fs_installer: Arc<dyn NpmPackageFsInstaller> =
-      match maybe_node_modules_path {
-        Some(node_modules_folder) => Arc::new(LocalNpmPackageInstaller::new(
-          lifecycle_scripts_executor,
-          npm_cache.clone(),
-          Arc::new(NpmPackageExtraInfoProvider::new(
-            npm_cache,
-            npm_registry_info_provider,
-            workspace_patch_packages,
-          )),
-          npm_install_deps_provider.clone(),
-          progress_bar.clone(),
-          npm_resolution.clone(),
-          sys,
-          tarball_cache,
-          node_modules_folder,
-          lifecycle_scripts,
-          system_info,
-        )),
-        None => Arc::new(GlobalNpmPackageInstaller::new(
-          npm_cache,
-          tarball_cache,
-          npm_resolution.clone(),
-          lifecycle_scripts,
-          system_info,
-        )),
-      };
-    Self {
-      fs_installer,
-      npm_install_deps_provider,
-      npm_resolution,
-      npm_resolution_initializer,
-      npm_resolution_installer,
-      maybe_lockfile,
-      top_level_install_flag: Default::default(),
-      cached_reqs: Default::default(),
-    }
-  }
-
-  /// Adds package requirements to the resolver and ensures everything is setup.
-  /// This includes setting up the `node_modules` directory, if applicable.
-  pub async fn add_and_cache_package_reqs(
-    &self,
-    packages: &[PackageReq],
-  ) -> Result<(), JsErrorBox> {
-    self.npm_resolution_initializer.ensure_initialized().await?;
-    self
-      .add_package_reqs_raw(
-        packages,
-        Some(PackageCaching::Only(packages.into())),
-      )
-      .await
-      .dependencies_result
-  }
-
-  pub async fn add_package_reqs_no_cache(
-    &self,
-    packages: &[PackageReq],
-  ) -> Result<(), JsErrorBox> {
-    self.npm_resolution_initializer.ensure_initialized().await?;
-    self
-      .add_package_reqs_raw(packages, None)
-      .await
-      .dependencies_result
-  }
-
-  pub async fn add_package_reqs(
-    &self,
-    packages: &[PackageReq],
-    caching: PackageCaching<'_>,
-  ) -> Result<(), JsErrorBox> {
-    self
-      .add_package_reqs_raw(packages, Some(caching))
-      .await
-      .dependencies_result
-  }
-
-  pub async fn add_package_reqs_raw(
-    &self,
-    packages: &[PackageReq],
-    caching: Option<PackageCaching<'_>>,
-  ) -> AddPkgReqsResult {
-    if packages.is_empty() {
-      return AddPkgReqsResult {
-        dependencies_result: Ok(()),
-        results: vec![],
-      };
-    }
-
-    #[cfg(debug_assertions)]
-    self.npm_resolution_initializer.debug_assert_initialized();
-
-    let mut result = self
-      .npm_resolution_installer
-      .add_package_reqs(packages)
-      .await;
-
-    if result.dependencies_result.is_ok() {
-      if let Some(lockfile) = self.maybe_lockfile.as_ref() {
-        result.dependencies_result = lockfile.error_if_changed();
-      }
-    }
-    if result.dependencies_result.is_ok() {
-      if let Some(caching) = caching {
-        // the async mutex is unfortunate, but needed to handle the edge case where two workers
-        // try to cache the same package at the same time. we need to hold the lock while we cache
-        // and since that crosses an await point, we need the async mutex.
-        //
-        // should have a negligible perf impact because acquiring the lock is still in the order of nanoseconds
-        // while caching typically takes micro or milli seconds.
-        let mut cached_reqs = self.cached_reqs.lock().await;
-        let uncached = {
-          packages
-            .iter()
-            .filter(|req| !cached_reqs.contains(req))
-            .collect::<Vec<_>>()
-        };
-
-        if !uncached.is_empty() {
-          result.dependencies_result = self.cache_packages(caching).await;
-          if result.dependencies_result.is_ok() {
-            for req in uncached {
-              cached_reqs.insert(req.clone());
-            }
-          }
-        }
-      }
-    }
-
-    result
-  }
-
-  pub async fn inject_synthetic_types_node_package(
-    &self,
-  ) -> Result<(), JsErrorBox> {
-    self.npm_resolution_initializer.ensure_initialized().await?;
-
-    // don't inject this if it's already been added
-    if self
-      .npm_resolution
-      .any_top_level_package(|id| id.nv.name == "@types/node")
-    {
-      return Ok(());
-    }
-
-    let reqs = &[PackageReq::from_str("@types/node").unwrap()];
-    self
-      .add_package_reqs(reqs, PackageCaching::Only(reqs.into()))
-      .await?;
-
-    Ok(())
-  }
-
-  pub async fn cache_package_info(
-    &self,
-    package_name: &str,
-  ) -> Result<Arc<NpmPackageInfo>, NpmRegistryPackageInfoLoadError> {
-    self
-      .npm_resolution_installer
-      .cache_package_info(package_name)
-      .await
-  }
-
-  pub async fn cache_packages(
-    &self,
-    caching: PackageCaching<'_>,
-  ) -> Result<(), JsErrorBox> {
-    self.npm_resolution_initializer.ensure_initialized().await?;
-    self.fs_installer.cache_packages(caching).await
-  }
-
-  pub fn ensure_no_pkg_json_dep_errors(
-    &self,
-  ) -> Result<(), Box<PackageJsonDepValueParseWithLocationError>> {
-    for err in self.npm_install_deps_provider.pkg_json_dep_errors() {
-      match err.source.as_kind() {
-        deno_package_json::PackageJsonDepValueParseErrorKind::VersionReq(_) => {
-          return Err(Box::new(err.clone()));
-        }
-        deno_package_json::PackageJsonDepValueParseErrorKind::Unsupported {
-          ..
-        } => {
-          // only warn for this one
-          log::warn!(
-            "{} {}\n    at {}",
-            colors::yellow("Warning"),
-            err.source,
-            err.location,
-          )
-        }
-      }
-    }
-    Ok(())
-  }
-
-  /// Ensures that the top level `package.json` dependencies are installed.
-  /// This may set up the `node_modules` directory.
-  ///
-  /// Returns `true` if the top level packages are already installed. A
-  /// return value of `false` means that new packages were added to the NPM resolution.
-  pub async fn ensure_top_level_package_json_install(
-    &self,
-  ) -> Result<bool, JsErrorBox> {
-    if !self.top_level_install_flag.raise() {
-      return Ok(true); // already did this
-    }
-
-    self.npm_resolution_initializer.ensure_initialized().await?;
-
-    let pkg_json_remote_pkgs = self.npm_install_deps_provider.remote_pkgs();
-    if pkg_json_remote_pkgs.is_empty() {
-      return Ok(true);
-    }
-
-    // check if something needs resolving before bothering to load all
-    // the package information (which is slow)
-    if pkg_json_remote_pkgs.iter().all(|pkg| {
-      self
-        .npm_resolution
-        .resolve_pkg_id_from_pkg_req(&pkg.req)
-        .is_ok()
-    }) {
-      log::debug!(
-        "All package.json deps resolvable. Skipping top level install."
-      );
-      return Ok(true); // everything is already resolvable
-    }
-
-    let pkg_reqs = pkg_json_remote_pkgs
-      .iter()
-      .map(|pkg| pkg.req.clone())
-      .collect::<Vec<_>>();
-    self.add_package_reqs_no_cache(&pkg_reqs).await?;
-
-    Ok(false)
-  }
-}
