+++ cli/util/mod.rs
@@ -4,7 +4,6 @@
 pub mod archive;
 pub mod collections;
 pub mod console;
-pub mod diff;
 pub mod display;
 pub mod draw_thread;
 pub mod extract;
