+++ cli/util/sync/mod.rs
@@ -1,9 +1,6 @@
 // Copyright 2018-2025 the Deno authors. MIT license.
 
 mod async_flag;
-mod task_queue;
 
 pub use async_flag::AsyncFlag;
 pub use deno_core::unsync::sync::AtomicFlag;
-pub use task_queue::TaskQueue;
-pub use task_queue::TaskQueuePermit;
