+++ resolvers/deno/Cargo.toml
@@ -28,6 +28,7 @@ deno_cache_dir.workspace = true
 deno_config.workspace = true
 deno_error.workspace = true
 deno_graph = { workspace = true, optional = true }
+deno_lockfile.workspace = true
 deno_media_type.workspace = true
 deno_npm.workspace = true
 deno_package_json.workspace = true
@@ -35,6 +36,7 @@ deno_path_util.workspace = true
 deno_semver.workspace = true
 deno_terminal.workspace = true
 deno_unsync.workspace = true
+dissimilar.workspace = true
 futures.workspace = true
 import_map.workspace = true
 indexmap.workspace = true
