+++ resolvers/npm_installer/Cargo.toml
@@ -21,11 +21,12 @@ boxed_error.workspace = true
 capacity_builder.workspace = true
 deno_config.workspace = true
 deno_error.workspace = true
+deno_lockfile.workspace = true
 deno_npm.workspace = true
 deno_npm_cache.workspace = true
 deno_package_json.workspace = true
 deno_path_util.workspace = true
-deno_resolver.workspace = true
+deno_resolver = { workspace = true, features = ["sync"] }
 deno_semver.workspace = true
 deno_terminal.workspace = true
 deno_unsync.workspace = true
